refactor: VideoProcessingService 분리
